requirements.txt
appnope==0.1.3
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
blis==0.7.7
cachetools==5.0.0
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
cymem==2.0.6
decorator==5.1.1
distro==1.7.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
executing==0.8.3
fonttools==4.32.0
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud-vision==2.7.2
googleapis-common-protos==1.56.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
install==1.3.5
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
kiwisolver==1.4.2
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
murmurhash==1.0.6
mypy==0.942
mypy-extensions==0.4.3
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
parso==0.8.3
pathspec==0.9.0
pathy==0.6.1
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
preshed==3.0.6
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.20.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
Pygments==2.11.2
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
receipt-parser-core==0.2.4
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
six==1.16.0
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
stack-data==0.2.0
tabula-py==2.3.0
tabulate==0.8.9
terminaltables==3.1.10
thinc==8.0.15
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
traitlets==5.1.1
typer==0.4.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
Wand==0.6.7
wasabi==0.9.1
wcwidth==0.2.5
